requirements.txt
accelerate>=0.27.0
diffusers>=0.27.0
transformers>=4.36.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.10.0
matplotlib>=3.7.0
wandb>=0.16.0
tensorboard>=2.15.0
datasets>=2.16.0
huggingface_hub>=0.20.0
